work on database class clean up
