+++ src/teehr/queries/duckdb.py
@@ -14,6 +14,8 @@ from teehr.models.queries import (
     TimeseriesCharQuery,
 )
 
+import teehr.queries.joined as tqj
+import teehr.queries.metrics as tqm
 import teehr.queries.utils as tqu
 import teehr.models.queries as tmq
 
@@ -53,7 +55,7 @@ def get_metrics(
         Must provide at least one.
     include_metrics : List[str]
         List of metrics (see below) for allowable list, or "all" to return all
-        Placeholder, currently ignored -> returns "all".
+        Must provide at least one.
     filters : Union[List[dict], None] = None
         List of dictionaries describing the "where" clause to limit data that
         is included in metrics.
@@ -158,86 +160,11 @@ def get_metrics(
         }
     )
 
-    query = f"""
-        WITH initial_joined AS (
-            SELECT
-                sf.reference_time
-                , sf.value_time as value_time
-                , sf.location_id as secondary_location_id
-                , pf.reference_time as primary_reference_time
-                , sf.value as secondary_value
-                , sf.configuration
-                , sf.measurement_unit
-                , sf.variable_name
-                , pf.value as primary_value
-                , pf.location_id as primary_location_id
-            FROM read_parquet({tqu._format_filepath(mq.secondary_filepath)}) sf
-            JOIN read_parquet({tqu._format_filepath(mq.crosswalk_filepath)}) cf
-                on cf.secondary_location_id = sf.location_id
-            JOIN read_parquet({tqu._format_filepath(mq.primary_filepath)}) pf
-                on cf.primary_location_id = pf.location_id
-                and sf.value_time = pf.value_time
-                and sf.measurement_unit = pf.measurement_unit
-                and sf.variable_name = pf.variable_name
-            {tqu.filters_to_sql(mq.filters)}
-        ),
-        joined AS (
-            {tqu._remove_duplicates_mq_cte(mq)}
-        )
-        {tqu._nse_cte(mq)}
-        {tqu._annual_metrics_cte(mq)}
-        {tqu._spearman_ranks_cte(mq)}
-        , metrics AS (
-            SELECT
-                {",".join([f"joined.{gb}" for gb in mq.group_by])}
-                {tqu._select_primary_count(mq)}
-                {tqu._select_secondary_count(mq)}
-                {tqu._select_primary_minimum(mq)}
-                {tqu._select_secondary_minimum(mq)}
-                {tqu._select_primary_maximum(mq)}
-                {tqu._select_secondary_maximum(mq)}
-                {tqu._select_primary_average(mq)}
-                {tqu._select_secondary_average(mq)}
-                {tqu._select_primary_sum(mq)}
-                {tqu._select_secondary_sum(mq)}
-                {tqu._select_primary_variance(mq)}
-                {tqu._select_secondary_variance(mq)}
-                {tqu._select_max_value_delta(mq)}
-                {tqu._select_mean_error(mq)}
-                {tqu._select_nash_sutcliffe_efficiency(mq)}
-                {tqu._select_nash_sutcliffe_efficiency_normalized(mq)}
-                {tqu._select_kling_gupta_efficiency(mq)}
-                {tqu._select_kling_gupta_efficiency_mod1(mq)}
-                {tqu._select_kling_gupta_efficiency_mod2(mq)}
-                {tqu._select_mean_absolute_error(mq)}
-                {tqu._select_mean_squared_error(mq)}
-                {tqu._select_root_mean_squared_error(mq)}
-                {tqu._select_primary_max_value_time(mq)}
-                {tqu._select_secondary_max_value_time(mq)}
-                {tqu._select_max_value_timedelta(mq)}
-                {tqu._select_relative_bias(mq)}
-                {tqu._select_multiplicative_bias(mq)}
-                {tqu._select_mean_absolute_relative_error(mq)}
-                {tqu._select_pearson_correlation(mq)}
-                {tqu._select_r_squared(mq)}
-                {tqu._select_spearman_correlation(mq)}
-            FROM
-                joined
-                {tqu._join_nse_cte(mq)}
-                {tqu._join_spearman_ranks_cte(mq)}
-            GROUP BY
-                {",".join([f"joined.{gb}" for gb in mq.group_by])}
-        )
-        SELECT
-            {",".join([f"metrics.{ob}" for ob in mq.group_by])}
-            {tqu.metrics_select_clause(mq)}
-            {tqu.geometry_select_clause(mq)}
-        FROM metrics
-            {tqu.metric_geometry_join_clause(mq)}
-            {tqu._join_annual_metrics_cte(mq)}
-        ORDER BY
-            {",".join([f"metrics.{ob}" for ob in mq.order_by])}
-    ;"""
+    select_joined_clause = tqj.select_joined_clause(mq.remove_duplicates)
+    joined = tqj.get_ind_parq_joined_timeseries_cte(mq, select_joined_clause)
+    metrics = tqm.get_metrics_clause(mq, tqu.geometry_joined_join_clause(mq))
+
+    query = joined + metrics
 
     if mq.return_query:
         return tqu.remove_empty_lines(query)
@@ -343,41 +270,21 @@ def get_joined_timeseries(
         }
     )
 
-    query = f"""
-        WITH initial_joined as (
-            SELECT
-                sf.reference_time,
-                sf.value_time,
-                sf.location_id as secondary_location_id,
-                sf.value as secondary_value,
-                sf.configuration,
-                sf.measurement_unit,
-                sf.variable_name,
-                pf.reference_time as primary_reference_time,
-                pf.value as primary_value,
-                pf.location_id as primary_location_id
-                {tqu.geometry_select_clause(jtq)}
-            FROM read_parquet({tqu._format_filepath(jtq.secondary_filepath)}) sf
-            JOIN read_parquet({tqu._format_filepath(jtq.crosswalk_filepath)}) cf
-                on cf.secondary_location_id = sf.location_id
-            JOIN read_parquet({tqu._format_filepath(jtq.primary_filepath)}) pf
-                on cf.primary_location_id = pf.location_id
-                and sf.value_time = pf.value_time
-                and sf.measurement_unit = pf.measurement_unit
-                and sf.variable_name = pf.variable_name
-            {tqu.geometry_join_clause(jtq)}
-            {tqu.filters_to_sql(jtq.filters)}
-        ),
-        joined AS (
-            {tqu._remove_duplicates_jtq_cte(jtq)}
-        )
+    select_joined_clause = tqj.select_joined_clause(jtq.remove_duplicates)
+    joined = tqj.get_ind_parq_joined_timeseries_cte(jtq, select_joined_clause)
+
+    select = f"""
         SELECT
-            *
+            joined.*
+            {tqu.geometry_select_clause(jtq)}
         FROM
             joined
+            {tqu.geometry_joined_join_clause(jtq)}
         ORDER BY
             {",".join(jtq.order_by)}
     ;"""
+
+    query = joined + select
 
     if jtq.return_query:
         return tqu.remove_empty_lines(query)
@@ -559,7 +466,7 @@ def get_timeseries_chars(
         }
     )
 
-    join_max_time_on = tqu._join_time_on(
+    join_max_time_on = tqu.join_time_on(
         join="mxt", join_to="chars", join_on=tcq.group_by
     )
